Add input `no-qt-binaries` as a synonym for `tools-only`
If a user wants to install source, docs, or examples, without installing
Qt binaries, then that user would have to turn on `tools-only`.
That option does not make semantic sense for these use cases, since the
user is not installing tools alone, and may not be installing tools at
all.

This change allows a user to turn off installation of Qt binaries using
an option that makes sense semantically.